+++ gamma_dashboard/static/dashboard/js/app/src/components/leaderboard/LeaderboardTable.jsx
@@ -1,12 +1,12 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 
-import LeaderboardTableRow from './LeadeboardTableRow';
+import LeaderboardCard from './LeaderboardCard';
 import { Loader } from '../generic';
 
 import { getUserStatus } from '../../utils';
 import { useTranslate } from '../../i18n/utils';
-import { ProgressPropType, StatusPropType } from '../propTypes';
+import { ProfilePropType, StatusPropType } from '../propTypes';
 
 const LeaderboardTable = ({
   rank, profiles, systemStatuses, delimiter,
@@ -14,19 +14,19 @@ const LeaderboardTable = ({
   const translatedEmptyText = useTranslate('leaderboard.status.empty.text');
 
   return (
-    <div className="LeaderboardTable" data-testid="leaderboard-table">
-      {!profiles ? (
+    <div className="leaderboard-table" data-testid="leaderboard-table">
+      {!profiles.length ? (
         <Loader />
       ) : (
         profiles.map((profile, index) => (
           <React.Fragment key={profile.userUid}>
-            <LeaderboardTableRow
-              rank={rank}
+            <LeaderboardCard
+              rank={rank || 0}
               profile={profile}
               status={getUserStatus(systemStatuses, profile.points, translatedEmptyText)}
             />
             {index === delimiter && (
-              <div className="LeaderboardTable-Separator" data-testid="leaderboard-table-separator">
+              <div className="leaderboard-table-separator" data-testid="leaderboard-table-separator">
                 <span /> <span /> <span />
               </div>
             )}
@@ -38,28 +38,16 @@ const LeaderboardTable = ({
 };
 
 LeaderboardTable.propTypes = {
-  profiles: PropTypes.arrayOf(
-    PropTypes.shape({
-      user_uid: PropTypes.string,
-      signup_source: PropTypes.string,
-      points: PropTypes.number,
-      badges: PropTypes.objectOf(
-        PropTypes.shape({
-          title: PropTypes.string,
-          description: PropTypes.string,
-          done: PropTypes.bool,
-          progress: PropTypes.shape(ProgressPropType),
-          url: PropTypes.string,
-        }),
-      ),
-    }),
-  ).isRequired,
-  systemStatuses: PropTypes.arrayOf(PropTypes.shape(StatusPropType)).isRequired,
-  rank: PropTypes.number.isRequired,
+  profiles: PropTypes.arrayOf(PropTypes.shape(ProfilePropType)),
+  systemStatuses: PropTypes.arrayOf(PropTypes.shape(StatusPropType)),
+  rank: PropTypes.number,
   delimiter: PropTypes.number,
 };
 
 LeaderboardTable.defaultProps = {
+  profiles: [],
+  systemStatuses: [],
+  rank: 0,
   delimiter: null,
 };