+++ gamma_dashboard/static/dashboard/js/app/src/components/leaderboard/__tests__/Badge.test.jsx
@@ -1,23 +1,18 @@
 import React from 'react';
 
 import '@testing-library/jest-dom';
-import { render, cleanup } from '@testing-library/react';
+import { cleanup } from '@testing-library/react';
 
 import Badge from '../Badge';
+import { renderWithProviders } from '../../../setupTests';
 
 afterEach(cleanup);
 
 describe('<Badge>', () => {
-  it('renders', () => {
-    const { getByTestId } = render(<Badge />);
-
-    expect(getByTestId('leaderboard-badge')).toBeInTheDocument();
-  });
-
   it('has correct src attribute', () => {
     const testUrl = 'https://localhost/static/images/link.png';
 
-    const { getByTestId } = render(<Badge url={testUrl} />);
+    const { getByTestId } = renderWithProviders(<Badge url={testUrl} />);
     const badgeImage = getByTestId('leaderboard-badge');
 
     expect(badgeImage).toHaveAttribute('src', testUrl);