+++ gamma_dashboard/static/dashboard/js/app/src/components/leaderboard/__tests__/utils.test.js
@@ -1,5 +1,5 @@
-import DataLeaderboardPage from './__mock__/DataLeaderboardPage.json';
 import { getLeaderboardTableProps } from '../utils';
+import DataLeaderboardPage from './__mock__/DataLeaderboardPage.json';
 
 describe('getLeaderboardTableProps', () => {
   it.each(DataLeaderboardPage)('should return the correct props for given parameters', ({ state, expectedProps }) => {